Document auto-generated slot accessors.
This forces (setf-s) documentation to return the respective
documentation for the function corresponding to slot accessor.

This is not particularly an erroneous behavior if the accessor is
undocumented, but it eases the burden of NASDF undocumented exports
checking, where we're using too much hacks to discover slot
accessors. And it's nicer to have some documentation to accessor than
to have none.